pyproject.toml
[project]
name = "whisper-ui"
version = "0.1.0"
description = "Speech-to-text system using OpenAI Whisper with Streamlit UI, speaker diarization, and Docker GPU support"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.0,<3",
  "pydantic-settings>=2.0,<3",
  "redis>=5.0,<6",
  "rq>=1.16,<2",
]

[project.optional-dependencies]
frontend = ["streamlit>=1.33,<2"]
worker = [
  "whisperx>=3.8,<4",
  "pyannote.audio>=3.1,<=3.3.2",
  "opencc-python-reimplemented>=0.1.7",
  "python-docx>=1.0,<2",
  "torch>=2.0",
]
dev = ["whisper-ui[frontend]", "pytest>=8.0,<9", "pytest-cov>=5.0,<6"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["RUF012"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
